Refactor dashboard and deployment pages
Refactor the dashboard and deployment pages, moving relevant files into the `features` directory and updating import paths.

* **Dashboard Refactor:**
  - Move `Dashboard` component from `src/pages/console/Dashboard.jsx` to `src/features/dashboard/pages/Dashboard.jsx`.
  - Update import paths in `src/App.jsx` and `src/routes/index.jsx` to reflect the new location of the `Dashboard` component.
  - Add new file `src/features/dashboard/pages/Dashboard.jsx` with the contents of the `Dashboard` component.

* **Deployment Refactor:**
  - Move `Deployment` component from `src/pages/console/Deployment.jsx` to `src/features/deployment/pages/Deployment.jsx`.
  - Update import paths in `src/routes/index.jsx` to reflect the new location of the `Deployment` component.
  - Add new file `src/features/deployment/pages/Deployment.jsx` with the contents of the `Deployment` component.
  - Refactor the `Deployment` component to improve code readability and maintainability.

* **File Deletions:**
  - Delete `src/pages/console/Dashboard.jsx`.
  - Delete `src/pages/console/Deployment.jsx`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/lion-agi/lion-studio?shareId=XXXX-XXXX-XXXX-XXXX).
