+++ src/routes/index.jsx
@@ -5,9 +5,9 @@ import Login from '../pages/Login';
 import Registration from '../pages/Registration';
 import AuthCallback from '../pages/AuthCallback';
 import EmailConfirmation from '../pages/EmailConfirmation';
-import Dashboard from '../pages/console/Dashboard';
+import Dashboard from '@/features/dashboard/pages/Dashboard';
 import Integrations from '../pages/console/Integrations';
-import Deployment from '../pages/console/Deployment';
+import Deployment from '@/features/deployment/pages/Deployment';
 import Library from '../pages/console/library/Library';
 import WorkflowEditor from '../features/workflow/components/WorkflowEditor';
 import About from '../pages/About';
